requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
idna==3.6
MarkupSafe==2.1.5
pip==23.2.1
psycopg2==2.9.9
PyQt6==6.6.1
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
python-dotenv==1.0.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.31.0
setuptools==68.2.0
urllib3==2.2.1
wheel==0.41.2
zipp==3.18.1